Chrome: Refactor PostVisibility component and split it into reusable components
