+++ editor/sidebar/post-visibility/test/index.js
@@ -18,16 +18,16 @@ describe( 'PostVisibility', () => {
 	};
 
 	it( 'should not render the edit link if the user doesn\'t have the right capability', () => {
-		let wrapper = shallow( <PostVisibility visibility="public" user={ {} } /> );
+		let wrapper = shallow( <PostVisibility user={ {} } /> );
 		expect( wrapper.find( 'a' ) ).toHaveLength( 0 );
-		wrapper = shallow( <PostVisibility visibility="public" postType="post" user={
+		wrapper = shallow( <PostVisibility postType="post" user={
 			{ data: { capabilities: { publish_posts: false } } }
 		} /> );
 		expect( wrapper.find( 'Dropdown' ) ).toHaveLength( 0 );
 	} );
 
 	it( 'should render if the user has the correct capability', () => {
-		const wrapper = shallow( <PostVisibility visibility="public" user={ user } /> );
+		const wrapper = shallow( <PostVisibility user={ user } /> );
 		expect( wrapper.find( 'Dropdown' ) ).not.toHaveLength( 0 );
 	} );
 } );
